Completed Problem 2... honestly it is really slow to process because of the wordstops and the use of Rabin Karp... maybe should ask dr. to fully use Tries algorithm???
I created a class calles Cities.py in Problem2 folder and use it in the main program. I use Rabin Karp and Tries algorithm for the string Matching.

if you want to run this, run Problem_2.py . Also, delete all the HTML file first(though it is optional. Dont have to)